Refactor extension foldering and add skeletons for dynamics
