+++ src/components/extensions/common/CommonExtensionModal.tsx
@@ -2,8 +2,8 @@ import { SendMessageParams } from '@/services/subsocial/commentIds'
 import { useMessageData } from '@/stores/message'
 import { cx } from '@/utils/class-names'
 import { useEffect } from 'react'
-import ChatForm, { ChatFormProps } from '../chats/ChatForm'
-import Modal, { ModalProps } from '../modals/Modal'
+import ChatForm, { ChatFormProps } from '../../chats/ChatForm'
+import Modal, { ModalProps } from '../../modals/Modal'
 
 export type BeforeMessageResult = {
   newMessageParams?: SendMessageParams
