+++ src/components/extensions/donate/DonateMessagePreview.tsx
@@ -8,7 +8,7 @@ import { DonateProperies } from '@subsocial/api/types'
 import BigNumber from 'bignumber.js'
 import { formatUnits } from 'ethers'
 import { HiArrowUpRight } from 'react-icons/hi2'
-import CommonChatItem from '../CommonChatItem'
+import CommonChatItem from '../common/CommonChatItem'
 import { ExtensionChatItemProps } from '../types'
 import { getMessageExtensionProperties } from '../utils'
 import { explorerByChainName } from './api/config'
