+++ src/components/extensions/donate/DonateModal/DonateForm.tsx
@@ -20,7 +20,7 @@ import Image from 'next/image'
 import { useEffect, useState } from 'react'
 import urlJoin from 'url-join'
 import { useAccount, useNetwork } from 'wagmi'
-import CommonExtensionModal from '../../CommonExtensionModal'
+import CommonExtensionModal from '../../common/CommonExtensionModal'
 import { chainIdByChainName } from '../api/config'
 import { useDonate, useGetBalance } from '../api/hooks'
 import AmountInput from './AmountInput'
