+++ src/components/extensions/index.tsx
@@ -9,7 +9,6 @@ const ImageModal = dynamic(() => import('./image/ImageModal'), {
 const NftModal = dynamic(() => import('./nft/NftModal'), {
   ssr: false,
 })
-
 const SecretBoxModal = dynamic(() => import('./secret-box/SecretBoxModal'), {
   ssr: false,
 })
