+++ src/components/extensions/nft/NftChatItem.tsx
@@ -8,7 +8,7 @@ import { cx } from '@/utils/class-names'
 import truncate from 'lodash.truncate'
 import { useMemo } from 'react'
 import { useInView } from 'react-intersection-observer'
-import CommonChatItem from '../CommonChatItem'
+import CommonChatItem from '../common/CommonChatItem'
 import { ExtensionChatItemProps } from '../types'
 import { getMessageExtensionProperties } from '../utils'
 import { getMarketplaceFromLink } from './utils'
