+++ src/components/extensions/secret-box/SecretBoxChatItem.tsx
@@ -4,7 +4,7 @@ import useGetTheme from '@/hooks/useGetTheme'
 import { useMyAccount } from '@/stores/my-account'
 import { cx } from '@/utils/class-names'
 import { useEffect, useState } from 'react'
-import CommonChatItem from '../CommonChatItem'
+import CommonChatItem from '../common/CommonChatItem'
 import { ExtensionChatItemProps } from '../types'
 import { getMessageExtensionProperties } from '../utils'
 import { useDecodeSecretBox } from './utils'
